Fix six standalone defects found while auditing the remaining packages
- Scan the offhand in PlayerHazardHandler. It walked the main inventory and
  the armour slots but never inventory.offhand, so a radioactive item held in
  the left hand irradiated nobody - while the geiger counted it, since
  PlayerHandler.getInventoryRadiation does include getOffhandItem.
- Stop RBMKRodItem.readTag from writing to the stack. It is reached from
  read-only paths - the hazard modifiers and the tooltip - so the server wrote
  CUSTOM_DATA into a rod merely because it sat in an inventory, costing a slot
  sync and a re-save, and the client did the same while rendering a tooltip,
  leaving its copy different from the server's. Callers that actually change
  state go through saveTag.
- Make the radiation maps in PlayerHandler concurrent. The server thread
  writes them up to forty times per tick per player, while the client render
  thread reads through ParticleEffectClient; in singleplayer that is one
  process, and a plain HashMap can spin forever inside getNode when a read
  races a put.
- Store the geiger cache timestamp with putLong to match the getLong that
  reads it. The int cast wraps negative once gameTime passes
  Integer.MAX_VALUE, which would disable the cache permanently.
- Play the armour equip sound server-side only. PlatformHooks.playSound
  broadcasts without excluding the actor, so the client played it a second
  time on top.
- Wrap the meltdown in try/finally so RBMKColumnBlock.dropLids is always
  restored. An exception mid-meltdown left the global flag cleared for the
  rest of the process and RBMK lids stopped dropping when dismantled.

+++ src/main/java/com/hbm_m/powerarmor/ModPowerArmorItem.java
@@ -374,7 +374,9 @@ public class ModPowerArmorItem extends ModArmorFSBPowered {
         if (currentTick - lastCheckTick > 20) {
             boolean hasDevice = playerHasGeigerDevice(player);
             data.putBoolean(TAG_HAS_GEIGER_DEVICE, hasDevice);
-            data.putInt(TAG_GEIGER_CHECK_TICK, (int) currentTick);
+            // putLong под getLong выше: с putInt после gameTime > Integer.MAX_VALUE каст
+            // уходил в отрицательные значения, и кэш отключался навсегда.
+            data.putLong(TAG_GEIGER_CHECK_TICK, currentTick);
             return hasDevice;
         }
